[dp] represent the array length as an integer

+++ src/dp-omp/main.cpp
@@ -39,10 +39,10 @@ int main(int argc, char **argv)
   // rounded up to the nearest multiple of the LocalWorkSize
   int szGlobalWorkSize = shrRoundUp((int)szLocalWorkSize, iNumElements);  
 
-  const size_t src_size = szGlobalWorkSize * 4;
+  const int src_size = szGlobalWorkSize * 4;
   const size_t src_size_bytes = src_size * sizeof(float);
 
-  const size_t dst_size = szGlobalWorkSize;
+  const int dst_size = szGlobalWorkSize;
   const size_t dst_size_bytes = dst_size * sizeof(float);
 
   // Allocate and initialize host arrays
@@ -54,7 +54,7 @@ int main(int argc, char **argv)
   shrFillArray(srcB, 4 * iNumElements);
 
   printf("Global Work Size \t\t= %d\nLocal Work Size \t\t= %d\n# of Work Groups \t\t= %d\n\n", 
-           szGlobalWorkSize, szLocalWorkSize, (szGlobalWorkSize % szLocalWorkSize + szGlobalWorkSize/szLocalWorkSize)); 
+         szGlobalWorkSize, szLocalWorkSize, (szGlobalWorkSize % szLocalWorkSize + szGlobalWorkSize/szLocalWorkSize)); 
 
   #pragma omp target data map(to: srcA[0:src_size], srcB[0:src_size]) \
                           map(from: dst[0:dst_size])
